pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-code-review-agent"
version = "0.1.0"
description = "AI agent that reviews GitHub PRs for bugs, security issues, and code quality"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "langfuse>=2.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
